refactor :refactor: restructure files, add parseAndValidate, consolidate game logic
