+++ src/App.js
@@ -1,21 +1,19 @@
-import { input, print } from './IoHandller.js';
-import { validateInput } from './Validate.js';
-import { parseInputs } from './Parsing.js';
-import { initGame } from './Game.js';
-import { initRandomName } from './Random.js';
+import { input, print } from './handler/IoHandller.js';
+import { initGame } from './game/Game.js';
+import { parseAndValidate } from './parsers/ParseAndValidate.js';
 
 class App {
   async run() {
     try {
       const carInput = await input('경주할 자동차의 이름을 입력한다 : ');
       const countInput = await input('시도할 횟수를 입력한다 : ');
-      const parseCarInput = parseInputs(carInput);
-      validateInput({ car: parseCarInput, count: countInput });
-      const carNamesWithRandom = initRandomName(...parseCarInput);
-      initGame({ car: carNamesWithRandom, count: countInput });
+      const parsedCarInput = parseAndValidate({
+        car: carInput,
+        count: countInput,
+      });
+      initGame({ car: parsedCarInput, count: countInput });
     } catch (error) {
       print(error.message);
-      throw error;
     }
   }
 }
